requirements.txt
# Core Google Cloud dependencies
google-cloud-aiplatform>=1.38.0
google-cloud-logging>=3.8.0
google-cloud-secret-manager>=2.16.0
google-cloud-storage>=2.10.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23

# Domain and network tools
python-whois>=0.8.0
requests>=2.31.0

# Data processing
python-dotenv>=1.0.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1

# Code quality
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.1

# ADK (Agent Development Kit) - Real ADK for Phase 1 migration
google-adk>=0.1.0
google-generativeai>=0.3.0
google-genai>=0.2.0
opentelemetry-instrumentation-google-genai
